Day 52: 주식 거래 설정 프론트엔드 완성 + KIS API 키 UI + 도움말 개선
[Frontend - 신규 파일]
- StockTradingSettingsView.vue: 주식 거래 설정 페이지 (기술적 지표/매수매도/리스크 관리)
- api/stock.ts: 주식 전용 API 모듈 (stockInfoApi, stockSettingApi)
- types/stock.ts: StockInfo, StockTradingSettings 타입 정의

[Frontend - 수정 파일]
- router/index.ts: /stock-settings 라우트 추가
- TheSidebar.vue: 주식 거래 설정 메뉴 활성화 + 토글 버그 수정
- ProfileView.vue: KIS API 키 등록/관리 섹션 추가
- HelpView.vue: Phase 2 주식/ETF FAQ 섹션 추가

[Backend - 수정 파일]
- StockTradingSettingDTO.java: int → BigDecimal 타입 수정 (6개 필드)

[도움말 개선]
- 카드 단위 6개 → 지표 단위 16개 개별 도움말로 분리
- SVG 차트 4개 (MA/RSI/BB/트레일링 스톱)
- 비유 설명 12개 + 주식 기본값 반영
- Phase 2 전용: 레버리지 decay 설명, 최대 보유일수

테스트: 설정 저장/수정/삭제/초기화 전체 통과

+++ backend/src/main/java/com/cryptotrading/dto/stock/StockTradingSettingDTO.java
@@ -39,8 +39,8 @@ public class StockTradingSettingDTO {
     private Integer rsiBuyThreshold;
     private Integer rsiSellThreshold;
     private Integer bbPeriod;
-    private Integer bbMultiplier;
-    private Integer volumeThreshold;
+    private BigDecimal bbMultiplier;        // 볼린저밴드 배수
+    private BigDecimal volumeThreshold;     // 거래량 급증 기준(%)
     
     // 리스크 관리
     private Integer dailyTradeLimitPct;
